+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Update minor and patch dependencies
+
 ## [1.0.22] - 2024-04-12
 
 ### Added