+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 -   Update minor and patch dependencies
+-   Update `@testing-library/react` to v15
 
 ## [1.0.22] - 2024-04-12