+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.23] - 2024-04-18
+
 ### Added
 
 -   Implement `Breadcrumbs`, `StateSkeletonBar`, and `StateSkeletonCircular` core components